SIS/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
python-telegram-bot[asyncio]
python-dotenv
pydantic
httpx
python-slugify
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator
pypdf
python-docx
beautifulsoup4
python-multipart
